Optimized away a few float-integer conversions

+++ include/utility/utility.h
@@ -70,7 +70,7 @@ void            SKIF_Util_SetEffectivePowerModeNotifications (bool enable);
 
 bool            SKIF_Util_IsHDRSupported              (bool refresh = false);
 bool            SKIF_Util_IsHDRActive                 (bool refresh = false);
-int             SKIF_Util_GetSDRWhiteLevelForHMONITOR (HMONITOR hMonitor);
+float           SKIF_Util_GetSDRWhiteLevelForHMONITOR (HMONITOR hMonitor);
 bool            SKIF_Util_EnableHDROutput             (void);
 bool            SKIF_Util_RegisterHotKeyHDRToggle     (void);
 bool            SKIF_Util_UnregisterHotKeyHDRToggle   (void);
